+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to `Weasley` will be documented in this file.
 
 ## [0.7.0](https://github.com/rougin/weasley/compare/v0.6.4...v0.7.0) - Unreleased
 
+### Added
+- `Random` for generating random strings in `Assorted` directory
+
 ### Changed
 - Conformed HTTP middlewares from `http-interop/http-middleware` to `rougin/slytherin`'s own middleware
 - Improve code quality and code formatting with `phpstan`, `php-cs-fixer`